// src/app/constants/enums/rw.enum.ts
export enum RwIdEnum {
  SiKongZhaiXing = 1,
  DuGuQiuBai = 2,
  HuaManLou = 3,
  YeGuCheng = 4,
  LaoShiHeShang = 5,
  MuDaoRen = 6,
  YuLuoCha = 7,
  XiaoLaoTou = 8,
  GongJiu = 9,
  ShaMan = 10,
  XiaoLongNv = 11,
  GuoJing = 12,
  HuangRong = 13,
  HuangYaoShi = 14,
  ShenDiao = 15,
  ZhangSanFeng = 16,
  JinLunFaWang = 17,
  HeZuDao = 18,
  HuangShang = 19,
  TieShou = 20,
  ZhuiMing = 21,
  LengXie = 22,
  LiXunHuan = 23,
  ChuLiuXiang = 24,
  XiaoShiYiLang = 25,
  FengSiNiang = 26,
  WuMingLaoSeng = 27,
  ZhouBoTong = 28,
  LingHuChong = 29,
  RenYingYing = 30,
  ZhuGeXianSheng = 31,
  WangXiaoShi = 32,
  DuanZiYu = 33,
  SuYing = 34,
  SuRongRong = 35,

  // Special person below
  LuXiaoFeng = 101,
  XiMenChuiXue = 102,
  YangGuo = 103,
  GuoXiang = 104,
  WuQing = 105,

  // Not special person below
  WuShi = 301,
  DianXiaoEr = 302,
  LaoBanNiang = 303,
  LaoBan = 304,
  LaoHuLi = 305,
  ZhuoFeiFan = 306,
  YueBuQun = 307,
  FangZheng = 308,
  FangSheng = 309,
  HeShang = 310,
  JueYuan = 311,
  SunXiuQing = 312,
  GuSong = 313,
  YeKai = 314,
  YinXian = 315,
  BaiChouFei = 316,
  BaiYiRen = 317,
  LanYiRen = 318,
  WeiZiYun = 319,
  MiTe = 320,
  ZhuangJia = 321,
  KuZhu = 322,
  HeiYiRen = 323,

  NoBody = 1001,

  // miss person below
  // 岳不群 武师头像 307
  // 方证 308
  // 方生 方证头像 309
  // 和尚 方证头像 310
  // 觉远 方证头像 311
  // 孙秀青 老板娘头像 312
  // 孤松 313 卓非凡头像
  // 叶开 武师头像 314
  // 殷羡 武师头像 315
  // 白愁飞 武师头像 316
  // 白衣人 武师头像 317
  // 蓝衣人 武师头像 318
}